SonarQube Bug BalloonGame "Cast one of the operands of this addition operation to a "long""

+++ src/main/java/de/hs_rm/de/milefiz/game/model/minigames/BalloonGame.java
@@ -73,7 +73,7 @@ public class BalloonGame extends MiniGame {
 
         if (!timeoutStarted) {
             timeoutStarted = true;
-            scheduler.schedule(this::handleTimeout, getTimeOut() + INSTRUCTION_DELAY_SECONDS, TimeUnit.SECONDS);
+            scheduler.schedule(this::handleTimeout, getTimeOut() + (long) INSTRUCTION_DELAY_SECONDS, TimeUnit.SECONDS);
         }
     }
